+++ lib/src/db/lib_iso_express.dart
@@ -166,9 +166,9 @@ abstract class DailyIsoExpressReport extends IsoExpressReport {
     List<Map<String, dynamic>> data;
     try {
       data = processFile(file);
-      if (data.isEmpty) return Future.value(-1);
+      if (data.isEmpty) return await Future.value(-1);
       await dbConfig.coll.remove({'date': day.toString()});
-      return dbConfig.coll.insertAll(data).then((_) {
+      return await dbConfig.coll.insertAll(data).then((_) {
         print('--->  Inserted $reportName for day $day');
         return 0;
       }).catchError((e) {