+++ test/db/isoexpress/ncpc_rapid_response_report_test.dart
@@ -4,7 +4,7 @@ import 'package:elec_server/src/db/isoexpress/ncpc_rapid_response_pricing_report
 import 'package:timezone/standalone.dart';
 import 'package:timezone/timezone.dart';
 
-tests() async {
+Future<void> tests() async {
   var archive = NcpcRapidResponsePricingReportArchive();
   group('NCPC rapid response report', () {
     setUp(() async => await archive.dbConfig.db.open());
@@ -24,7 +24,7 @@ tests() async {
   });
 }
 
-insertDays() async {
+Future<void> insertDays() async {
   var archive = NcpcRapidResponsePricingReportArchive();
   await archive.dbConfig.db.open();
   var days = Interval(TZDateTime.utc(2017, 3, 2), TZDateTime.utc(2017, 4, 1))
@@ -36,7 +36,7 @@ insertDays() async {
   await archive.dbConfig.db.close();
 }
 
-main() async {
+Future<void> main() async {
 //  await NcpcRapidResponsePricingReportArchive().setupDb();
 
   await insertDays();