+++ test/db/weather/winter_storms_test.dart
@@ -58,7 +58,7 @@ Future<void> tests() async {
   // });
 }
 
-winterStormTests() async {
+Future<void> winterStormTests() async {
   var archive = WinterStormsArchive();
 
   await archive.setupDb();