+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Move ISONE Monthly Asset NCPC to DuckDB
 - test
 
+# 2026-08-17
+- Modify MIS report reading to allow for empty lines at the end of the CSV file
+- Fix several lints
+
 # 2026-08-14
 - Finish download_isone_ptid_table.dart job, send email, etc.